Change getting value from rapidjson so that it compiles under both: clang and gcc

+++ src/core/fields/cfield.cpp
@@ -38,16 +38,16 @@ mc::field::ptr mc::cfield::getPtr(const rapidjson::Value& json)
 		return field::ptr(new cfield(s, anyType::get(s, json["value"].Get<unsigned int>())));
 		break;
 	case mc::stype::int8:
-		return field::ptr(new cfield(s, anyType::get(s, json["value"].Get<long long>())));
+		return field::ptr(new cfield(s, anyType::get(s, json["value"].Get<int64_t>())));
 		break;
 	case mc::stype::uint8:
-		return field::ptr(new cfield(s, anyType::get(s, json["value"].Get<unsigned long long>())));
+		return field::ptr(new cfield(s, anyType::get(s, json["value"].Get<uint64_t>())));
 		break;
 	case mc::stype::int16:
-		return field::ptr(new cfield(s, anyType::get(s, json["value"].Get<long long>())));
+		return field::ptr(new cfield(s, anyType::get(s, json["value"].Get<int64_t>())));
 		break;
 	case mc::stype::uint16:
-		return field::ptr(new cfield(s, anyType::get(s, json["value"].Get<unsigned long long>())));
+		return field::ptr(new cfield(s, anyType::get(s, json["value"].Get<uint64_t>())));
 		break;
 	case mc::stype::float4:
 		return field::ptr(new cfield(s, anyType::get(s, json["value"].Get<float>())));
